+++ src/components/itemGuru/CLPManagement.vue
@@ -2,6 +2,7 @@
 import { onMounted, ref } from 'vue';
 import api from '../../services/api';
 import { useToast } from 'primevue';
+import Swal from 'sweetalert2';
 
 const toast = useToast()
 const allRpk = ref([])
@@ -27,6 +28,18 @@ const fetchRpk = async () => {
 }
 
 const deleteRpk = async (id) => {
+    const confirm = await Swal.fire({
+        title: "Are you sure, you want to delete this Learning Plan?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d33",
+        cancelButtonColor: "#3085d6",
+        confirmButtonText: "Yes, delete it",
+        cancelButtonText: "Cancel",
+    });
+
+    if (!confirm.isConfirmed) return;
+
     try {
         const res = await api.delete(`/rpk/${id}`)
         await fetchRpk()
